requirements.txt
fastapi
uvicorn[standard]
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-fastapi
